requirements_conda.txt
pytorch
pytorch-cpu
numpy
pillow
scikit-image
scikit-learn
torchfile
visdom
jsonpatch